edu_api2/requirement.txt
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.0.6
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-formtools==2.2
django-import-export==2.3.0
django-reversion==3.0.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
self==0.0.0
six==1.15.0
tablib==2.0.0
urllib3==1.25.10
vue==0.0.1
xadmin @ https://codeload.github.com/sshwsfc/xadmin/zip/django2
xlrd==1.2.0
xlwt==1.3.0
